Add a noRedirect authorization parameter.
Bumping the minor version number since this introduces a new capability which
*should* be backwards compatible.

The new functionality provided by this commit allows the redirect URL generated
by authorize() to be optionally returned in the promise instead of actually
performing the redirect. This can be useful if an application needs to perform
some actions after authorize() has been called but before the redirect actually
occurs. Testing and debugging is another use case for this change.

Tests all pass, which is expected since noRedirect already existed for testing
purposes; this commit just exposes it as an authorization parameter. If it is not
provided, it will be undefined and evaluate to falsey, which should help with
backward compatibility.